+++ paddle/phi/kernels/cpu/rnn_grad_kernel.cc
@@ -717,8 +717,8 @@ struct SingleGradLayer : GradLayer<T, GradCellType> {
     phi::funcs::SetConstant<CPUContext, T> zero;
     zero(dev_ctx, input_grad, static_cast<T>(0.0));
 
-    int64_t time_step = static_cast<int>(input->dims()[0]);
-    int64_t batch_size = static_cast<int>(input->dims()[1]);
+    int64_t time_step = input->dims()[0];
+    int64_t batch_size = input->dims()[1];
     int direction_num = is_bidirec ? 2 : 1;
 
     // in this section, create the gate_state_grad for the postprocess calculate
@@ -827,8 +827,8 @@ struct BidirGradLayer : GradLayer<T, GradCellType> {
                   int hidden_size,
                   const std::string& mode,
                   int gate_num) {
-    int64_t time_step = static_cast<int>(input->dims()[0]);
-    int64_t batch_size = static_cast<int>(input->dims()[1]);
+    int64_t time_step = input->dims()[0];
+    int64_t batch_size = input->dims()[1];
     int direction_num = is_bidirec ? 2 : 1;
     // split the output two tensor to output_forward, output_backward
     phi::funcs::SetConstant<CPUContext, T> zero;
